Deprecate galsim.integ.trapz and galsim.integ.midpt
